Done: Added task 2 to homework lesson 16
